+++ src/main/java/com/nbrichau/vanillaextension/init/FenceInit.java
@@ -39,6 +39,7 @@ public class FenceInit {
 	public static final Block gold_ore_fence = null;
 	public static final Block iron_ore_fence = null;
 	public static final Block coal_ore_fence = null;
+	public static final Block nether_gold_ore_fence = null;
 	public static final Block oak_log_fence = null;
 	public static final Block spruce_log_fence = null;
 	public static final Block birch_log_fence = null;
@@ -110,6 +111,7 @@ public class FenceInit {
 	public static final Block pumpkin_fence = null;
 	public static final Block netherrack_fence = null;
 	public static final Block soul_sand_fence = null;
+	public static final Block soul_soil_fence = null;
 	public static final Block glowstone_fence = null;
 	public static final Block carved_pumpkin_fence = null;
 	public static final Block jack_o_lantern_fence = null;
@@ -307,6 +309,7 @@ public class FenceInit {
 		event.getRegistry().register(new OreFence(Block.Properties.from(Blocks.GOLD_ORE)).setRegistryName("gold_ore_fence"));
 		event.getRegistry().register(new OreFence(Block.Properties.from(Blocks.IRON_ORE)).setRegistryName("iron_ore_fence"));
 		event.getRegistry().register(new OreFence(Block.Properties.from(Blocks.COAL_ORE)).setRegistryName("coal_ore_fence"));
+		event.getRegistry().register(new OreFence(Block.Properties.from(Blocks.NETHER_GOLD_ORE)).setRegistryName("nether_gold_ore_fence"));
 		event.getRegistry().register(new LogFence(Block.Properties.create(Material.WOOD, MaterialColor.WOOD).hardnessAndResistance(2.0F).sound(SoundType.WOOD)).setRegistryName("oak_log_fence"));
 		event.getRegistry().register(new LogFence(Block.Properties.create(Material.WOOD, MaterialColor.OBSIDIAN).hardnessAndResistance(2.0F).sound(SoundType.WOOD)).setRegistryName("spruce_log_fence"));
 		event.getRegistry().register(new LogFence(Block.Properties.create(Material.WOOD, MaterialColor.SAND).hardnessAndResistance(2.0F).sound(SoundType.WOOD)).setRegistryName("birch_log_fence"));
@@ -378,6 +381,7 @@ public class FenceInit {
 		event.getRegistry().register(new PumpkinFence(Block.Properties.from(Blocks.PUMPKIN)).setRegistryName("pumpkin_fence"));
 		event.getRegistry().register(new FenceBlock(Block.Properties.from(Blocks.NETHERRACK)).setRegistryName("netherrack_fence"));
 		event.getRegistry().register(new FenceBlock(Block.Properties.create(Material.SAND, MaterialColor.BROWN).hardnessAndResistance(0.5F).speedFactor(0.4F).sound(SoundType.SAND)).setRegistryName("soul_sand_fence"));
+		event.getRegistry().register(new FenceBlock(Block.Properties.from(Blocks.SOUL_SOIL)).setRegistryName("soul_soil_fence"));
 		event.getRegistry().register(new FenceBlock(Block.Properties.from(Blocks.GLOWSTONE)).setRegistryName("glowstone_fence"));
 		event.getRegistry().register(new FenceBlock(Block.Properties.from(Blocks.CARVED_PUMPKIN)).setRegistryName("carved_pumpkin_fence"));
 		event.getRegistry().register(new FenceBlock(Block.Properties.from(Blocks.JACK_O_LANTERN)).setRegistryName("jack_o_lantern_fence"));
@@ -576,6 +580,7 @@ public class FenceInit {
 		event.getRegistry().register(new BlockItem(gold_ore_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("gold_ore_fence"));
 		event.getRegistry().register(new BlockItem(iron_ore_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("iron_ore_fence"));
 		event.getRegistry().register(new BlockItem(coal_ore_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("coal_ore_fence"));
+		event.getRegistry().register(new BlockItem(nether_gold_ore_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("nether_gold_ore_fence"));
 		event.getRegistry().register(new BlockItem(oak_log_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("oak_log_fence"));
 		event.getRegistry().register(new BlockItem(spruce_log_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("spruce_log_fence"));
 		event.getRegistry().register(new BlockItem(birch_log_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("birch_log_fence"));
@@ -647,6 +652,7 @@ public class FenceInit {
 		event.getRegistry().register(new BlockItem(pumpkin_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("pumpkin_fence"));
 		event.getRegistry().register(new BlockItem(netherrack_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("netherrack_fence"));
 		event.getRegistry().register(new BlockItem(soul_sand_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("soul_sand_fence"));
+		event.getRegistry().register(new BlockItem(soul_soil_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("soul_soil_fence"));
 		event.getRegistry().register(new BlockItem(glowstone_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("glowstone_fence"));
 		event.getRegistry().register(new BlockItem(carved_pumpkin_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("carved_pumpkin_fence"));
 		event.getRegistry().register(new BlockItem(jack_o_lantern_fence, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("jack_o_lantern_fence"));