+++ src/main/java/com/nbrichau/vanillaextension/walls/OreWall.java
@@ -27,10 +27,9 @@ public class OreWall extends WallBlock {
 			return MathHelper.nextInt(rand, 2, 5);
 		} else if (this == WallInit.redstone_ore_wall) {
 			return MathHelper.nextInt(rand, 1, 5);
-		}/* else {
+		} else {
 			return this == WallInit.nether_gold_ore_wall ? MathHelper.nextInt(rand, 0, 1) : 0;
-		}*/
-		return 0;
+		}
 	}
 
 	@Override